remove `operator_` prefix on files in operatorapi
