+++ operatorapi/namespaces.go
