+++ operatorapi/namespaces_test.go
