+++ operatorapi/nodes_test.go
