+++ operatorapi/session.go
